+++ generator/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/ARM-software/embedded-development-services-client/generator/cmd"
+
+func main() {
+	cmd.Execute()
+}